Also move compact_level0_phase1
